# StudennykApi/openapi.yaml
openapi: 3.0.0
info:
  version: '1'
  title: Studennyk API
paths:
  /api/v1/tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: integer
                format: int64
                example: 123

  /api/v1/tasks/{task_id}:
    get:
      summary: Get a specific task by ID
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    put:
      summary: Update specific fields of a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Title:
                  type: string
                Description:
                  type: string
                Due:
                  type: string
                  format: date-time
                IsDone:
                  type: boolean
      responses:
        '204':
          description: Task fields updated successfully
        '404':
          description: Task not found

    delete:
      summary: Delete a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        Id:
          type: integer
          format: int64
          example: -1
        Title:
          type: string
          example: Task Title
        Description:
          type: string
          example: Task Description
        Created:
          type: string
          format: date-time
          example: '2023-01-01T00:00:00Z'
        Due:
          type: string
          format: date-time
          example: '2023-01-02T00:00:00Z'
        IsDone:
          type: boolean
          example: false